fix: re-rendering of child components due to timer
