+++ src/components/ResourceBrowser/ResourceList/ResourceList.tsx
@@ -4,8 +4,6 @@ import {
     convertToOptionsList,
     createGroupSelectList,
     filterImageList,
-    getTimeElapsed,
-    handleUTCTime,
     processK8SObjects,
     sortObjectArrayAlphabetically,
 } from '../../common'
@@ -43,7 +41,6 @@ import { CreateResource } from './CreateResource'
 import { AppDetailsTabs, AppDetailsTabsIdPrefix } from '../../v2/appDetails/appDetails.store'
 import NodeDetailComponent from '../../v2/appDetails/k8Resource/nodeDetail/NodeDetail.component'
 import { SelectedResourceType } from '../../v2/appDetails/appDetails.type'
-import moment from 'moment'
 import ConnectingToClusterState from './ConnectingToClusterState'
 import { SOME_ERROR_MSG } from '../../../config/constantMessaging'
 import searchWorker from '../../../config/searchWorker'
@@ -51,7 +48,6 @@ import WebWorker from '../../app/WebWorker'
 import { ShortcutProvider } from 'react-keybind'
 import { DynamicTabs, useTabs } from '../../common/DynamicTabs'
 import {
-    checkIfDataIsStale,
     getEventObjectTypeGVK,
     getGroupedK8sObjectMap,
     getK8SObjectMapAfterGroupHeadingClick,
@@ -72,7 +68,6 @@ import ClusterTerminal from '../../ClusterNodes/ClusterTerminal'
 import { createTaintsList } from '../../cluster/cluster.util'
 import NodeDetailsList from '../../ClusterNodes/NodeDetailsList'
 import NodeDetails from '../../ClusterNodes/NodeDetails'
-let interval
 
 
 export default function ResourceList() {
@@ -107,7 +102,6 @@ export default function ResourceList() {
     const [selectedResource, setSelectedResource] = useState<ApiResourceGroupType>(null)
     const [logSearchTerms, setLogSearchTerms] = useState<Record<string, string>>()
     const [lastDataSyncTimeString, setLastDataSyncTimeString] = useState('')
-    const [timeElapsedLastSync, setTimeElapsedLastSync] = useState('')
     const [lastDataSync, setLastDataSync] = useState(false)
     const [showCreateResourceModal, setShowCreateResourceModal] = useState(false)
     const [resourceSelectionData, setResourceSelectionData] = useState<Record<string, ApiResourceGroupType>>()
@@ -361,21 +355,7 @@ export default function ResourceList() {
         }
     }, [selectedNamespace])
 
-    useEffect(() => {
-        const _lastDataSyncTime = Date()
-        const _staleDataCheckTime = moment()
-        isStaleDataRef.current = false
-        setLastDataSyncTimeString(` ${handleUTCTime(_lastDataSyncTime, true)}`)
-        interval = setInterval(() => {
-            checkIfDataIsStale(isStaleDataRef, _staleDataCheckTime)
-            setLastDataSyncTimeString(` ${handleUTCTime(_lastDataSyncTime, true)}`)
-            setTimeElapsedLastSync(getTimeElapsed(_lastDataSyncTime, moment()))
-        }, 1000)
-        return () => {
-            setTimeElapsedLastSync('')
-            clearInterval(interval)
-        }
-    }, [lastDataSync])
+   
 
     const getDetailsClusterList = async () => {
         setTerminalLoader(true)
@@ -741,8 +721,6 @@ export default function ResourceList() {
     )
 
     const refreshData = (): void => {
-        clearInterval(interval)
-        setTimeElapsedLastSync('')
         setSelectedResource(null)
         getSidebarData(selectedCluster.value)
     }
@@ -1036,7 +1014,7 @@ export default function ResourceList() {
                     }}
                 >
                     <div className="resource-browser-tab flex left w-100">
-                        <DynamicTabs tabs={tabs} removeTabByIdentifier={removeTabByIdentifier} stopTabByIdentifier={stopTabByIdentifier} enableShortCut={!showCreateResourceModal} timeElapsedLastSync={timeElapsedLastSync} refreshData={refreshData} loader={loader||rawGVKLoader||clusterLoader||resourceListLoader} isOverview={isOverview}/>
+                        <DynamicTabs tabs={tabs} removeTabByIdentifier={removeTabByIdentifier} stopTabByIdentifier={stopTabByIdentifier} enableShortCut={!showCreateResourceModal} refreshData={refreshData} lastDataSync={lastDataSync} loader={loader||rawGVKLoader||clusterLoader||resourceListLoader} isOverview={isOverview} isStaleDataRef={isStaleDataRef} setLastDataSyncTimeString={setLastDataSyncTimeString}/>
                     </div>
                 </div>
                 {renderResourceBrowser()}
