+++ src/components/common/DynamicTabs/DynamicTabs.tsx
@@ -15,6 +15,11 @@ import { MoreButtonWrapper, noMatchingTabs, TabsMenu } from './DynamicTabs.compo
 import Select from 'react-select/dist/declarations/src/Select'
 import { AppDetailsTabs } from '../../../components/v2/appDetails/appDetails.store'
 import './DynamicTabs.scss'
+import moment from 'moment'
+import { handleUTCTime,getTimeElapsed } from '../helpers/time'
+import {checkIfDataIsStale
+} from '../../ResourceBrowser/Utils'
+let interval
 
 /**
  * This component enables a way to display dynamic tabs with the following functionalities,
@@ -31,10 +36,12 @@ function DynamicTabs({
     stopTabByIdentifier,
     enableShortCut,
     shortcut,
-    timeElapsedLastSync,
     refreshData,
     loader,
-    isOverview
+    isOverview,
+    lastDataSync,
+    setLastDataSyncTimeString,
+    isStaleDataRef
 }: DynamicTabsProps & IWithShortcut) {
     const { push } = useHistory()
     const tabsSectionRef = useRef<HTMLDivElement>(null)
@@ -48,6 +55,24 @@ function DynamicTabs({
     const [isMenuOpen, setIsMenuOpen] = useState(false)
     const tabPopupMenuRef = useRef(null)
     const CLUSTER_TERMINAL_TAB = 'cluster_terminal-Terminal'
+    const [timeElapsedLastSync, setTimeElapsedLastSync] = useState('')
+
+
+    useEffect(() => {
+        const _lastDataSyncTime = Date()
+        const _staleDataCheckTime = moment()
+        isStaleDataRef.current = false
+        setLastDataSyncTimeString(` ${handleUTCTime(_lastDataSyncTime, true)}`)
+        interval = setInterval(() => {
+            checkIfDataIsStale(isStaleDataRef, _staleDataCheckTime)
+            setLastDataSyncTimeString(` ${handleUTCTime(_lastDataSyncTime, true)}`)
+            setTimeElapsedLastSync(getTimeElapsed(_lastDataSyncTime, moment()))
+        }, 1000)
+        return () => {
+            setTimeElapsedLastSync('')
+            clearInterval(interval)
+        }
+    }, [lastDataSync])
 
     useEffect(() => {
         initTabsData(tabs, setTabsData, setSelectedTab, closeMenu)
@@ -281,7 +306,11 @@ function DynamicTabs({
                         <div>
                             <RefreshIcon
                                 className="icon-dim-16 scn-6 flexbox mr-6 cursor ml-12"
-                                onClick={refreshData}
+                                onClick={() => {
+                                    clearInterval(interval)
+                                    setTimeElapsedLastSync('')
+                                    refreshData()
+                                }}
                             />
                         </div>
                     </Tippy>
@@ -308,55 +337,53 @@ function DynamicTabs({
                     </ul>
                 </div>
             )}
-            <div className='ml-auto flexbox'>
-            {!isOverview && selectedTab.id !== CLUSTER_TERMINAL_TAB && (
-                <div className="flexbox dc__border-left fw-6 cn-7 dc__align-self-stretch dc__align-items-center dc__no-shrink">
-                    {timerForSync()}
-                </div>
-            )}
+            <div className="ml-auto flexbox dc__no-shrink dc__align-self-stretch dc__border-left">
+                {!isOverview && selectedTab?.id !== CLUSTER_TERMINAL_TAB && (
+                    <div className="flexbox fw-6 cn-7 dc__align-items-center">{timerForSync()}</div>
+                )}
 
-            {tabsData.dynamicTabs.length > 0 && (
-                <MoreButtonWrapper
-                    tabPopupMenuRef={tabPopupMenuRef}
-                    isMenuOpen={isMenuOpen}
-                    onClose={closeMenu}
-                    toggleMenu={toggleMenu}
-                >
-                    <div className="more-tabs__search-icon icon-dim-16 cursor-text" onClick={focusSearchTabInput}>
-                        <SearchIcon className="icon-dim-16" />
-                    </div>
-                    <ReactSelect
-                        ref={moreButtonRef}
-                        placeholder="Search tabs"
-                        classNamePrefix="tab-search-select"
-                        options={tabsData.dynamicTabs}
-                        value={selectedTab}
-                        inputValue={tabSearchText}
-                        onChange={onChangeTab}
-                        onKeyDown={escHandler}
-                        onInputChange={handleOnChangeSearchText}
-                        tabSelectsValue={false}
-                        backspaceRemovesValue={false}
-                        controlShouldRenderValue={false}
-                        hideSelectedOptions={false}
-                        menuIsOpen
-                        autoFocus
-                        noOptionsMessage={noMatchingTabs}
-                        components={{
-                            IndicatorSeparator: null,
-                            DropdownIndicator: null,
-                            Option: tabsOption,
-                            Menu: TabsMenu,
-                        }}
-                        styles={COMMON_TABS_SELECT_STYLES}
-                    />
-                    <div className="more-tabs__clear-tab-search icon-dim-16 cursor">
-                        {tabSearchText && (
-                            <ClearIcon className="clear-tab-search-icon icon-dim-16" onClick={clearSearchInput} />
-                        )}
-                    </div>
-                </MoreButtonWrapper>
-            )}
+                {tabsData.dynamicTabs.length > 0 && (
+                    <MoreButtonWrapper
+                        tabPopupMenuRef={tabPopupMenuRef}
+                        isMenuOpen={isMenuOpen}
+                        onClose={closeMenu}
+                        toggleMenu={toggleMenu}
+                    >
+                        <div className="more-tabs__search-icon icon-dim-16 cursor-text" onClick={focusSearchTabInput}>
+                            <SearchIcon className="icon-dim-16" />
+                        </div>
+                        <ReactSelect
+                            ref={moreButtonRef}
+                            placeholder="Search tabs"
+                            classNamePrefix="tab-search-select"
+                            options={tabsData.dynamicTabs}
+                            value={selectedTab}
+                            inputValue={tabSearchText}
+                            onChange={onChangeTab}
+                            onKeyDown={escHandler}
+                            onInputChange={handleOnChangeSearchText}
+                            tabSelectsValue={false}
+                            backspaceRemovesValue={false}
+                            controlShouldRenderValue={false}
+                            hideSelectedOptions={false}
+                            menuIsOpen
+                            autoFocus
+                            noOptionsMessage={noMatchingTabs}
+                            components={{
+                                IndicatorSeparator: null,
+                                DropdownIndicator: null,
+                                Option: tabsOption,
+                                Menu: TabsMenu,
+                            }}
+                            styles={COMMON_TABS_SELECT_STYLES}
+                        />
+                        <div className="more-tabs__clear-tab-search icon-dim-16 cursor">
+                            {tabSearchText && (
+                                <ClearIcon className="clear-tab-search-icon icon-dim-16" onClick={clearSearchInput} />
+                            )}
+                        </div>
+                    </MoreButtonWrapper>
+                )}
             </div>
         </div>
     )
