+++ src/components/common/DynamicTabs/Types.ts
@@ -26,10 +26,12 @@ export interface DynamicTabsProps {
     removeTabByIdentifier: (id: string) => string
     stopTabByIdentifier: (title: string) => string
     enableShortCut?: boolean
-    timeElapsedLastSync: string
     loader: boolean
     refreshData: () => void
-    isOverview: boolean
+    isOverview: boolean,
+    lastDataSync:boolean,
+    setLastDataSyncTimeString: (time: string) => void,
+    isStaleDataRef: any,
 }
 
 export interface TabsDataType {
